Added Service_CreateStream.py to app.yaml. Moved CreateStream post() method to Service_CreateStream.py. Added Javascript file to color active link in nav bar to the MasterTemplate.html.
