+++ templates/Create.html
@@ -6,7 +6,7 @@
     <p>
     Create New Stream:
     </p>
-    <form action="create" method="post">
+    <form action="/services/createstream" method="post">
         Stream Name:<br>
         <input type="text" name="streamname"><br>
         Subscriber Emails (semicolon delimited):<br>
