+++ templates/MasterTemplate.html
@@ -5,6 +5,7 @@
     <meta charset="UTF-8">
     <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css">
     <link rel="stylesheet" type="text/css" href="/css/baseStyle.css">
+    <script src="/js/bootstrap/jquery-slim.min.js"></script>
 
 
     <title>{% block title %}{% endblock %}</title>
@@ -24,7 +25,7 @@
         <hr>
 
         <div class="topnav">
-            <a href="manage" class="active">Manage</a>
+            <a href="manage">Manage</a>
             <a href="create">Create</a>
             <a href="view">View All</a>
             <a href="search">Search</a>
@@ -36,5 +37,8 @@
             {% endblock %}
         </div>
     </div>
+
+    <script src="/js/ActivePageColoring.js"></script>
+
 </body>
 </html>
